+++ src/Core/Network.h
@@ -55,7 +55,7 @@ namespace rdf {
 
 namespace net {
 
-	DllCoreExport bool urlExists(const QString& url, int port = 80);
+	DllCoreExport bool urlExists(const QString& url, quint16 port = 80);
 	DllCoreExport QByteArray download(const QString& url, bool* ok = 0);
 }
